% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_tutorial.R
\name{access_tutorial}
\alias{access_tutorial}
\title{access_tutorial}
\usage{
access_tutorial(tutorial)
}
\arguments{
\item{tutorial}{Name of tutorial you wish to access.}
}
\description{
A wrapper function to deploy a learnr tutorial in the ucrstats package.
}
