% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_tutorial.R
\name{available_tutorials}
\alias{available_tutorials}
\title{available_tutorials}
\usage{
available_tutorials()
}
\arguments{

}
\description{
A function that lists all of the available tutorials in the "ucrstats" package
}
