% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_tutorial.R
\name{list_tutorials}
\alias{list_tutorials}
\title{list_tutorials}
\usage{
list_tutorials()
}
\arguments{

}
\description{
A function that lists all of the available tutorials in the "ucrstats" package
}
