% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lme_one}
\alias{lme_one}
\title{A simulated data set used to demonstrate how to fit a linear mixed-effects model with lme4. There are 500 participants. Each participants has up to 25 repeated measurements recorded.}
\format{
A data frame with 6720 rows and 7 variables:
\describe{
\item{id}{An ID variable grouping the repeated measurements}
\item{time_id}{time point values}
\item{Y}{Outcome}
\item{X1}{Predictor Variable X1}
\item{X2}{Predictor Variable X2}
\item{X3}{Predictor Variable X3}
\item{X4}{Predictor Variable X4}
}
}
\usage{
lme_one
}
\description{
A simulated data set used to demonstrate how to fit a linear mixed-effects model with lme4. There are 500 participants. Each participants has up to 25 repeated measurements recorded.
}
\keyword{datasets}
