% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lme_three}
\alias{lme_three}
\title{A simulated data set used to demonstrate how to fit a linear mixed-effects model with lme4. There are 1000 participants. Each participants has up to 25 repeated measurements recorded.}
\format{
A data frame with 13643 rows and 8 variables:
\describe{
\item{id}{An ID variable grouping the repeated measurements}
\item{time_id}{time point values}
\item{Y}{Outcome}
\item{X1}{Predictor Variable X1}
\item{X2}{Predictor Variable X2}
\item{X3}{Predictor Variable X3}
\item{X4}{Predictor Variable X4}
\item{X5}{Predictor Variable X5}
}
}
\usage{
lme_three
}
\description{
A simulated data set used to demonstrate how to fit a linear mixed-effects model with lme4. There are 1000 participants. Each participants has up to 25 repeated measurements recorded.
}
\keyword{datasets}
