% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long}
\alias{long}
\title{A simulated data set used to represent long data.}
\format{
A data frame with 4000 rows and 5 variables:
\describe{
\item{ID1}{An ID variable grouping the repeated measurements}
\item{time}{The time point for each repeated measurement}
\item{mean}{mean of observation at time point}
\item{sd}{mean of observation at time point}
\item{median}{mean of observation at time point}
}
}
\usage{
long
}
\description{
A simulated data set used to represent long data.
}
\keyword{datasets}
