% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutorials}
\alias{tutorials}
\title{A table containing a list of all the tutorials in the package}
\format{
A data frame
\describe{
\item{ID}{The string placed in the \code{access_tutorial} function to access the tutorial}
\item{Subject}{The topic the tutorial generally relates to}
\item{Description}{A brief description of what the tutorial will cover}

}
}
\usage{
tutorials
}
\description{
A table containing a list of all the tutorials in the package
}
\keyword{datasets}
