% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wide}
\alias{wide}
\title{A simulated dataset used to represent wide data.}
\format{
A data frame with 1000 rows and 13 variables:
\describe{
\item{ID1}{An ID variable grouping the repeated measurements}
\item{v1/mean}{Variable representing the mean at time point 1}
\item{v1/median}{Variable representing the median at time point 1}
\item{v1/sd}{Variable representing the sd at time point 1}
\item{v2/mean}{Variable representing the mean at time point 2}
\item{v2/median}{Variable representing the median at time point 2}
\item{v2/sd}{Variable representing the sd at time point 2}
\item{v3/mean}{Variable representing the mean at time point 3}
\item{v3/median}{Variable representing the median at time point 3}
\item{v3/sd}{Variable representing the sd at time point 3}
\item{v4/mean}{Variable representing the mean at time point 4}
\item{v4/median}{Variable representing the median at time point 4}
\item{v4/sd}{Variable representing the sd at time point 4}
}
}
\usage{
wide
}
\description{
A simulated dataset used to represent wide data.
}
\keyword{datasets}
